Move and rename new remote as MfaRemote
